requirements.txt
moviepy>=1.0.3
openai-whisper>=20240930
torch>=2.5.0
torchaudio>=2.5.0
numpy>=1.26.0
Pillow>=10.0.0
pydub>=0.25.1
setuptools<81
